refactor: make search form a live component
